refactor: rewrite to app directory & update to next15
