+++ frontend/package.json
@@ -35,11 +35,11 @@
     "jotai": "^2.9.3",
     "lru-cache": "^11.0.1",
     "lucide-react": "^0.426.0",
-    "next": "^14.2.3",
+    "next": "^15.0.3",
     "next-sitemap": "^4.2.3",
     "oauth-1.0a": "^2.2.6",
-    "react": "^18",
-    "react-dom": "^18",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-icons": "^5.3.0",
     "sharp": "^0.33.5",
     "tailwind-merge": "^2.5.2",
@@ -56,6 +56,7 @@
     "@types/react-dom": "^18",
     "@typescript-eslint/parser": "^8.3.0",
     "eslint": "^8",
+    "eslint-config-next": "^15.0.3",
     "knip": "^5.27.4",
     "postcss": "^8",
     "postcss-load-config": "^6.0.1",
