+++ frontend/src/app/api/logout/route.ts
@@ -1,12 +1,13 @@
-import { cookies } from "next/headers";
+import { cookies as cookiesPromise } from "next/headers";
 import { redirect } from "next/navigation";
 
-export const GET = () => {
-  cookies().delete({
+export const GET = async () => {
+  const cookies = await cookiesPromise();
+  cookies.delete({
     name: "access_token",
     path: "/",
   });
-  cookies().delete({
+  cookies.delete({
     name: "access_token_secret",
     path: "/",
   });
