+++ frontend/src/app/api/profile/route.ts
@@ -4,24 +4,17 @@ import { createUsosService } from "@/lib/usos";
 
 export const GET = async () => {
   try {
-    await createUsosService().getProfile();
+    const usos = await createUsosService();
+    await usos.getProfile();
 
     return NextResponse.json(
-      {
-        msg: "Authorized",
-      },
-      {
-        status: 200,
-      },
+      { msg: "Authorized" },
+      { status: 200 }
     );
   } catch (e) {
     return NextResponse.json(
-      {
-        msg: "Unauthorized",
-      },
-      {
-        status: 401,
-      },
+      { msg: "Unauthorized" },
+      { status: 401 }
     );
   }
 };
