+++ frontend/src/app/plans/preview/[id]/_components/SharePlanPage.tsx
@@ -1,36 +1,21 @@
+"use client";
+
 import { useAtom } from "jotai";
-import type { GetServerSideProps, InferGetServerSidePropsType } from "next";
-import Link from "next/link";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import * as React from "react";
 import { LuDownloadCloud } from "react-icons/lu";
 
 import { planFamily } from "@/atoms/planFamily";
 import { plansIds } from "@/atoms/plansIds";
 import { ClassSchedule } from "@/components/ClassSchedule";
-import { SolvroLogo } from "@/components/SolvroLogo";
-import { Button, buttonVariants } from "@/components/ui/button";
+import { Button } from "@/components/ui/button";
 import { usePlan } from "@/lib/usePlan";
-import { cn } from "@/lib/utils";
 import { Day } from "@/services/usos/types";
 
-// eslint-disable-next-line @typescript-eslint/require-await
-export const getServerSideProps = (async (context) => {
-  const { id } = context.query;
-
-  if (typeof id !== "string") {
-    throw new Error(`Invalid hash ${id?.toString()}`);
-  }
-
-  return { props: { id } };
-}) satisfies GetServerSideProps;
-
-const SharePlan = ({
-  id,
-}: InferGetServerSidePropsType<typeof getServerSideProps>) => {
+export default function SharePlanPage({ planId }: { planId: string }) {
   const uuid = React.useMemo(() => crypto.randomUUID(), []);
   const [plans, setPlans] = useAtom(plansIds);
-  const plan = usePlan({ planId: id });
+  const plan = usePlan({ planId });
   const [planToCopy, setPlanToCopy] = useAtom(planFamily({ id: uuid }));
 
   const router = useRouter();
@@ -52,38 +37,11 @@ const SharePlan = ({
     });
 
     setTimeout(() => {
-      void router.push(`/app/createplan/${newPlan.id}`);
+      router.push(`/plans/create/${newPlan.id}`);
     }, 200);
   };
   return (
-    <div className="flex max-h-screen flex-col overflow-x-hidden">
-      <div className="flex max-h-20 min-h-20 items-center justify-between bg-mainbutton7">
-        <div className="ml-4 w-1/4 flex-none">
-          <SolvroLogo />
-        </div>
-        <div className="mr-4 flex items-center justify-end gap-4">
-          <Link
-            href="/app/plans"
-            data-umami-event="Back to plans"
-            className={cn(
-              buttonVariants({ variant: "link" }),
-              "p-0 text-white",
-            )}
-          >
-            <span className="text-nowrap">Moje plany</span>
-          </Link>
-          <Link
-            href="/"
-            data-umami-event="Back to homepage"
-            className={cn(
-              buttonVariants({ variant: "link" }),
-              "p-0 text-white",
-            )}
-          >
-            <span className="text-nowrap">Strona główna</span>
-          </Link>
-        </div>
-      </div>
+    <>
       <div className="flex items-center justify-center gap-4 p-2">
         <Button
           onClick={copyPlan}
@@ -130,19 +88,6 @@ const SharePlan = ({
             ),
         )}
       </div>
-      <div className="flex w-full flex-1 items-center justify-center bg-mainbutton7 p-2">
-        <p className="text-center text-white">
-          Made with ❤️ by{" "}
-          <a
-            href="https://solvro.pwr.edu.pl/"
-            className="font-bold text-mainbutton hover:underline"
-          >
-            SOLVRO
-          </a>
-        </p>
-      </div>
-    </div>
+    </>
   );
-};
-
-export default SharePlan;
+}
