+++ frontend/src/components/PlanDisplayLink.tsx
@@ -7,7 +7,7 @@ import { buttonVariants } from "./ui/button";
 export function PlanDisplayLink({ id }: { id: string }) {
   return (
     <Link
-      href={`/app/preview/${id}`}
+      href={`/plans/preview/${id}`}
       className={cn(buttonVariants(), "font-semibold")}
     >
       <span>Podgląd</span>
