+++ frontend/src/components/footer.tsx
@@ -1,0 +1,15 @@
+export const Footer = () => {
+  return (
+    <footer className="sm:mt-12">
+      <p className="p-4 text-center text-white">
+        Made with ❤️ by{" "}
+        <a
+          href="https://solvro.pwr.edu.pl/"
+          className="font-bold text-mainbutton hover:underline"
+        >
+          SOLVRO
+        </a>
+      </p>
+    </footer>
+  );
+};
